Factor messy llm.py file and make "valid models" optional
Summary:
This will factor the LLM file into two seperate files, one for the parallel processing of the prompts that instruct/autocomplete/mitre use, and the baseline abstractions for openai/anyscale/together (the file that we expect open source owners to extend when testing their own LLMs).

I also remove the "requirement" that models are in the valid model list (as this is more of a precaution - but there's several scenarios where new models are available on these platforms and folks will just want to test them without altering code).

Reviewed By: dkgi

Differential Revision: D51926877

fbshipit-source-id: 6d06559bf641fe8f58ee6819d4489bdc2cdf9fcd